Re-organise routes and controllers
