+++ controllers/auth/auth.controller.js
@@ -1,33 +1,6 @@
 const model = require('../../models/index');
 const bcrypt = require('bcryptjs');
 
-exports.findAll = (req, res) => {
-  model.User.findAll()
-    .then(users => {
-      res.send(users);
-    })
-    .catch(err => {
-      res.status(500).send({
-        message:
-          err.message || "Some error occurred while retrieving users."
-      });
-    });
-}
-
-exports.findOne = (req, res) => {
-  model.User.findByPk(req.params.id)
-    .then(user => {
-      if (user == null) {
-        res.status(404).send({ message: `User with id: ${req.params.id} does not exist` });
-      } else {
-        res.send(user);
-      }
-    })
-    .catch(err => {
-      res.status(500).send({ message: err.message || 'Something went wrong' });
-    })
-}
-
 exports.login = (request, response) => {
 	let { external_id, password } =  request.body;
   let hash = bcrypt.hashSync(password, 10);
